Add psycopg2 to Pipfile for SQLAlchemy Postgresql support
I'm not sure why this isn't installed automatically, but I guess in order to make sure tons of unused dependencies aren't installed on machines that don't need them as well as all the errors that crop up, it wasn't added.

As a user who's simply using it for deployment and didn't need it for local development, I suppose this is what happens.